fix(test): UBSAN signed-integer overflow in TimerMeasuresElapsedTime; update stage docs and issue descriptions
- tests/unit/test_PassContext.cpp: change `volatile int sum` to
  `volatile long long` — sum(0..99999) = 4,999,950,000 which exceeds
  INT_MAX; UBSAN with halt_on_error=1 was aborting the ASAN CI job

- docs/phases/stage-1-ir-pass-manager/: fill in exit-report.md, kickoff.md;
  mark all todo.md items complete; update spec.md DoD to clang-18 and ✅

- docs/phases/stage-2-synthesis/: write spec.md, todo.md, kickoff.md for
  Stage 2 (SynthesisProvider Concept, GridSynthProvider, SKProvider,
  TGateSynthesisPass); remove empty misnamed stage-2-ir/ directory

- memory-bank/: advance activeContext to Stage 2 planning; mark Stage 1
  complete in progress.md with 2026-04-25 completion date

- CHANGELOG.md: add UBSAN overflow and gcc-9.4 debug preset to
  Failed Approaches; mark Stage 1 progress log entries complete

+++ tests/unit/test_PassContext.cpp
@@ -65,7 +65,8 @@ TEST(PassContext, TimerMeasuresElapsedTime) {
     PassContext ctx{5};
     ctx.startTimer("test");
     // Minimal spin so the timer actually ticks
-    volatile int sum = 0;
+    // Use long long to avoid signed-integer overflow under UBSAN (sum of 0..99999 > INT_MAX)
+    volatile long long sum = 0;
     for (int i = 0; i < 100000; ++i) sum += i;
     (void)sum;
     ctx.stopTimer();
